Move StatusBarNotification.getDetails to NotificationDetails.kt
